web_api/requirements.txt
starlette
uvicorn
httpx==0.12.1
tortoise-orm
asyncpg==0.20.1
pydantic==1.5.1
cassandra-driver==3.14.0
cqlsh==5.0.4
cqlengine==0.21.0
faust
faust[uvloop,fast,rocksdb,redis]
asyncio
asgiref
syncer

numpy
scipy
pandas